pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["animal_classification*"]
namespaces = false

[tool.setuptools.package-data]
"animal_classification" = ["data/mini_animals/**/*"]

[project]
name = "animal-classification-aldanondo-malon"
version = "1.1.0"
description = "A deep learning project to classify 90 animals using VGG architectures."
readme = "README.md"
requires-python = ">=3.10,<=3.13"
license = "MIT"

authors = [
{ name = "Andrés Malón", email = "andresmalon@gmail.com" },
{ name = "Roberto Aldanondo", email = "robertoaj02@gmail.com" }
]

dependencies = [
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "torchaudio>=2.8.0",
    "lightning>=2.5.5",
    "pillow>=10.0.0",
    "kagglehub>=0.2.0",
    "gradio>=4.44.1",
    "numpy>=2.0.2",
    "pandas>=2.3.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.9.4",
    "scikit-learn>=1.6.1",
    "twine"
]

[project.optional-dependencies]
dev = ["pdoc>=15.0.4"]
cli = ["kagglehub"]

[project.urls]
Homepage = "https://a139573.github.io/animal-classification/"
Repository = "https://github.com/a139573/animal-classification"

[project.scripts]
animal-train = "animal_classification.modeling.train:main"
animal-infer = "animal_classification.modeling.inference:main"
animal-dashboard = "animal_classification.app.dashboard:main"
animal-download-data = "animal_classification.preprocessing.download_data:main"
animal-reduce-data = "animal_classification.preprocessing.reduce_data:main"

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cuda-128"
url = "https://download.pytorch.org/whl/cu128" 
explicit = true

[[tool.uv.index]]
name = "pytorch-cuda-126"
url = "https://download.pytorch.org/whl/cu126" 
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cuda-128" },
]
torchvision = [
    { index = "pytorch-cuda-126"},
]
torchaudio = [
    { index = "pytorch-cuda-126" },
]

[tool.pdoc]
docformat = "numpy"